refactor: generalize quantized speculative verification
