+++ mlx_vlm/models/glm5_next/exact_ops.py
@@ -529,171 +529,6 @@ if (matrix_row < HC) {
 """
 
 
-_AFFINE5_HEADER = _COMMON_HEADER + r"""
-inline void decode_affine5(
-    const device uint8_t* w,
-    float scale,
-    float bias,
-    thread float* out) {
-  uint8_t w0 = w[0];
-  uint8_t w1 = w[1];
-  uint8_t w2 = w[2];
-  uint8_t w3 = w[3];
-  uint8_t w4 = w[4];
-  out[0] = (w0 & 0x1f) * scale + bias;
-  out[1] = (((w0 & 0xe0) >> 5) + ((w1 & 0x03) << 3)) * scale + bias;
-  out[2] = ((w1 & 0x7c) >> 2) * scale + bias;
-  out[3] = (((w1 & 0x80) >> 7) + ((w2 & 0x0f) << 1)) * scale + bias;
-  out[4] = (((w2 & 0xf0) >> 4) + ((w3 & 0x01) << 4)) * scale + bias;
-  out[5] = ((w3 & 0x3e) >> 1) * scale + bias;
-  out[6] = (((w3 & 0xc0) >> 6) + ((w4 & 0x07) << 2)) * scale + bias;
-  out[7] = ((w4 & 0xf8) >> 3) * scale + bias;
-}
-"""
-
-
-_AFFINE4_HEADER = _COMMON_HEADER + r"""
-inline void decode_affine4(
-    const device uint8_t* w,
-    float scale,
-    float bias,
-    thread float* out) {
-  uint8_t w0 = w[0];
-  uint8_t w1 = w[1];
-  uint8_t w2 = w[2];
-  uint8_t w3 = w[3];
-  out[0] = (w0 & 0x0f) * scale + bias;
-  out[1] = ((w0 & 0xf0) >> 4) * scale + bias;
-  out[2] = (w1 & 0x0f) * scale + bias;
-  out[3] = ((w1 & 0xf0) >> 4) * scale + bias;
-  out[4] = (w2 & 0x0f) * scale + bias;
-  out[5] = ((w2 & 0xf0) >> 4) * scale + bias;
-  out[6] = (w3 & 0x0f) * scale + bias;
-  out[7] = ((w3 & 0xf0) >> 4) * scale + bias;
-}
-"""
-
-
-_AFFINE5_WIDE_SOURCE = r"""
-uint tile = threadgroup_position_in_grid.y;
-uint simd_gid = simdgroup_index_in_threadgroup;
-uint simd_lid = thread_index_in_simdgroup;
-
-constexpr int K_LANES = 8;
-constexpr int ROWS_PER_SIMD = 32 / K_LANES;
-constexpr int ROWS_PER_TG = 2 * ROWS_PER_SIMD;
-constexpr int SUB = 8;
-
-short k_lane = simd_lid % K_LANES;
-short sg_row = simd_lid / K_LANES;
-int out_row = int(tile) * ROWS_PER_TG + int(simd_gid) * ROWS_PER_SIMD + sg_row;
-int row = min(out_row, N_SIZE - 1);
-
-constexpr int W_ROW_BYTES = K_SIZE * 5 / 8;
-constexpr int GROUPS = K_SIZE / GROUP_SIZE;
-const device uint8_t* wrow = (const device uint8_t*)w + row * W_ROW_BYTES;
-const device T* srow = scales + row * GROUPS;
-const device T* brow = biases + row * GROUPS;
-
-float result[VERIFY_T][BATCH_SIZE];
-for (int t = 0; t < VERIFY_T; ++t) {
-  for (int b = 0; b < BATCH_SIZE; ++b) {
-    result[t][b] = 0.0f;
-  }
-}
-
-for (int g = k_lane; g < GROUPS; g += K_LANES) {
-  float scale = float(srow[g]);
-  float bias = float(brow[g]);
-  for (int sc = 0; sc < GROUP_SIZE / SUB; ++sc) {
-    int k0 = g * GROUP_SIZE + sc * SUB;
-    const device uint8_t* wc = wrow + k0 * 5 / 8;
-    float wdq[SUB];
-    decode_affine5(wc, scale, bias, wdq);
-    for (int t = 0; t < VERIFY_T; ++t) {
-      for (int b = 0; b < BATCH_SIZE; ++b) {
-        const device T* xc = x + (b * VERIFY_T + t) * K_SIZE + k0;
-        float accum = 0.0f;
-        for (int i = 0; i < SUB; ++i) {
-          accum += float(xc[i]) * wdq[i];
-        }
-        result[t][b] += accum;
-      }
-    }
-  }
-}
-
-for (int t = 0; t < VERIFY_T; ++t) {
-  for (int b = 0; b < BATCH_SIZE; ++b) {
-    result[t][b] += simd_shuffle_down(result[t][b], 4);
-    result[t][b] += simd_shuffle_down(result[t][b], 2);
-    result[t][b] += simd_shuffle_down(result[t][b], 1);
-  }
-}
-
-if (k_lane == 0 && out_row < N_SIZE) {
-  for (int t = 0; t < VERIFY_T; ++t) {
-    for (int b = 0; b < BATCH_SIZE; ++b) {
-      y[(b * VERIFY_T + t) * N_SIZE + out_row] = T(result[t][b]);
-    }
-  }
-}
-"""
-
-
-_AFFINE5_WIDE_ARGMAX_SOURCE = _AFFINE5_WIDE_SOURCE.replace(
-    r"""if (k_lane == 0 && out_row < N_SIZE) {
-  for (int t = 0; t < VERIFY_T; ++t) {
-    for (int b = 0; b < BATCH_SIZE; ++b) {
-      y[(b * VERIFY_T + t) * N_SIZE + out_row] = T(result[t][b]);
-    }
-  }
-}
-""",
-    r"""threadgroup float simd_best_values[VERIFY_T][BATCH_SIZE][2];
-threadgroup int simd_best_indices[VERIFY_T][BATCH_SIZE][2];
-
-for (int t = 0; t < VERIFY_T; ++t) {
-  for (int b = 0; b < BATCH_SIZE; ++b) {
-    float rounded = float(T(result[t][b]));
-    float best = simd_shuffle(rounded, 0);
-    int best_index = int(tile) * ROWS_PER_TG + int(simd_gid) * ROWS_PER_SIMD;
-    for (int row = 1; row < ROWS_PER_SIMD; ++row) {
-      float candidate = simd_shuffle(rounded, row * K_LANES);
-      int candidate_index =
-          int(tile) * ROWS_PER_TG + int(simd_gid) * ROWS_PER_SIMD + row;
-      if (candidate > best) {
-        best = candidate;
-        best_index = candidate_index;
-      }
-    }
-    if (simd_lid == 0) {
-      simd_best_values[t][b][simd_gid] = best;
-      simd_best_indices[t][b][simd_gid] = best_index;
-    }
-  }
-}
-threadgroup_barrier(mem_flags::mem_threadgroup);
-
-if (simd_gid == 0 && simd_lid == 0) {
-  for (int t = 0; t < VERIFY_T; ++t) {
-    for (int b = 0; b < BATCH_SIZE; ++b) {
-      float best = simd_best_values[t][b][0];
-      int best_index = simd_best_indices[t][b][0];
-      if (simd_best_values[t][b][1] > best) {
-        best = simd_best_values[t][b][1];
-        best_index = simd_best_indices[t][b][1];
-      }
-      int offset = (b * VERIFY_T + t) * NUM_TILES + int(tile);
-      tile_values[offset] = T(best);
-      tile_indices[offset] = best_index;
-    }
-  }
-}
-""",
-)
-
-
 _AFFINE5_SWITCH_HEADER = _COMMON_HEADER + r"""
 constant constexpr int SIMD_SIZE = 32;
 constant constexpr int PACK_FACTOR = 8;
@@ -911,75 +746,6 @@ for (int row = 0; row < RESULTS_PER_SIMDGROUP; ++row) {
 """
 
 
-_AFFINE5_DENSE_BLOCK_SOURCE = r"""
-uint n_tile = threadgroup_position_in_grid.y;
-uint simd_gid = simdgroup_index_in_threadgroup;
-uint simd_lid = thread_index_in_simdgroup;
-
-int out_row = int(n_tile) * ROWS_PER_TG +
-    int(simd_gid) * RESULTS_PER_SIMDGROUP;
-constexpr int W_ROW_BYTES = K_SIZE * 5 / 8;
-constexpr int GROUPS = K_SIZE / GROUP_SIZE;
-
-const device uint8_t* ws = (const device uint8_t*)w +
-    out_row * W_ROW_BYTES + int(simd_lid) * PACKS_PER_THREAD * BYTES_PER_PACK;
-const device T* sc = scales + out_row * GROUPS +
-    int(simd_lid) / SCALE_STEP_PER_THREAD;
-const device T* bs = biases + out_row * GROUPS +
-    int(simd_lid) / SCALE_STEP_PER_THREAD;
-
-float result[VERIFY_T][RESULTS_PER_SIMDGROUP] = {0.0f};
-float x_thread[VERIFY_T][VALUES_PER_THREAD];
-const device T* xk = x + int(simd_lid) * VALUES_PER_THREAD;
-
-for (int k = 0; k < K_SIZE; k += BLOCK_SIZE) {
-  float sums[VERIFY_T];
-  bool active = k + int(simd_lid) * VALUES_PER_THREAD < K_SIZE;
-  for (int token = 0; token < VERIFY_T; ++token) {
-    if (active) {
-      sums[token] = load_affine5_vector_exact<T>(
-          xk + token * K_SIZE,
-          x_thread[token]);
-    } else {
-      sums[token] = 0.0f;
-      for (int i = 0; i < VALUES_PER_THREAD; ++i) {
-        x_thread[token][i] = 0.0f;
-      }
-    }
-  }
-  for (int row = 0; row < RESULTS_PER_SIMDGROUP; ++row) {
-    const device uint8_t* wr = ws + row * W_ROW_BYTES;
-    const device T* sr = sc + row * GROUPS;
-    const device T* br = bs + row * GROUPS;
-    for (int token = 0; token < VERIFY_T; ++token) {
-      if (active) {
-        result[token][row] += affine5_qdot_exact(
-            wr,
-            x_thread[token],
-            float(sr[0]),
-            float(br[0]),
-            sums[token]);
-      }
-    }
-  }
-  ws += BLOCK_SIZE * 5 / 8;
-  sc += BLOCK_SIZE / GROUP_SIZE;
-  bs += BLOCK_SIZE / GROUP_SIZE;
-  xk += BLOCK_SIZE;
-}
-
-for (int token = 0; token < VERIFY_T; ++token) {
-  for (int row = 0; row < RESULTS_PER_SIMDGROUP; ++row) {
-    int n = out_row + row;
-    float value = simd_sum(result[token][row]);
-    if (simd_lid == 0 && n < N_SIZE) {
-      y[token * N_SIZE + n] = T(value);
-    }
-  }
-}
-"""
-
-
 _AFFINE5_MOE_DOWN_SOURCE = r"""
 uint n_tile = threadgroup_position_in_grid.y;
 uint token = threadgroup_position_in_grid.z;
@@ -1064,158 +830,6 @@ for (int row = 0; row < RESULTS_PER_SIMDGROUP; ++row) {
   float shared_value = simd_sum(shared_result[row]);
   if (simd_lid == 0 && n < N_SIZE) {
     y[int(token) * N_SIZE + n] = T(routed_sum[row] + T(shared_value));
-  }
-}
-"""
-
-
-_MXFP8_HEADER = _COMMON_HEADER + r"""
-struct fp8_e4m3 {
-  operator half() {
-    uint16_t value = (bits & 127) << 7;
-    half converted = as_type<half>(value);
-    converted *= 256.0h;
-    return (bits & 128) ? -converted : converted;
-  }
-  operator float() { return float(this->operator half()); }
-  uint8_t bits;
-};
-
-inline float decode_e4m3(uint8_t value) {
-  return float(*(thread fp8_e4m3*)(&value));
-}
-
-inline float decode_e8m0(uint8_t value) {
-  uint32_t output = value == 0 ? 0x400000 : (uint32_t(value) << 23);
-  return as_type<float>(output);
-}
-"""
-
-
-_MXFP8_WIDE_SOURCE = r"""
-uint tile = threadgroup_position_in_grid.y;
-uint simd_gid = simdgroup_index_in_threadgroup;
-uint simd_lid = thread_index_in_simdgroup;
-
-constexpr int K_LANES = 16;
-constexpr int ROWS_PER_SIMD = 32 / K_LANES;
-constexpr int ROWS_PER_TG = 2 * ROWS_PER_SIMD;
-constexpr int NF4 = GROUP_SIZE / 4;
-
-short k_lane = simd_lid % K_LANES;
-short sg_row = simd_lid / K_LANES;
-int out_row = int(tile) * ROWS_PER_TG + int(simd_gid) * ROWS_PER_SIMD + sg_row;
-int row = min(out_row, N_SIZE - 1);
-
-constexpr int GROUPS = K_SIZE / GROUP_SIZE;
-const device uint8_t* wrow = (const device uint8_t*)w + row * K_SIZE;
-const device uint8_t* srow = scales + row * GROUPS;
-
-float result[VERIFY_T][BATCH_SIZE];
-for (int t = 0; t < VERIFY_T; ++t) {
-  for (int b = 0; b < BATCH_SIZE; ++b) {
-    result[t][b] = 0.0f;
-  }
-}
-
-for (int g = k_lane; g < GROUPS; g += K_LANES) {
-  int k0 = g * GROUP_SIZE;
-  float scale = decode_e8m0(srow[g]);
-  const device uint8_t* wg = wrow + k0;
-  float4 w4[NF4];
-  for (int i = 0; i < NF4; ++i) {
-    w4[i] = float4(
-        decode_e4m3(wg[4 * i]),
-        decode_e4m3(wg[4 * i + 1]),
-        decode_e4m3(wg[4 * i + 2]),
-        decode_e4m3(wg[4 * i + 3]));
-  }
-  for (int t = 0; t < VERIFY_T; ++t) {
-    for (int b = 0; b < BATCH_SIZE; ++b) {
-      const device vec<T, 4>* xv4 =
-          (const device vec<T, 4>*)(x + (b * VERIFY_T + t) * K_SIZE + k0);
-      float accum = 0.0f;
-      for (int i = 0; i < NF4; ++i) {
-        accum += dot(w4[i], float4(xv4[i]));
-      }
-      result[t][b] += scale * accum;
-    }
-  }
-}
-
-for (int t = 0; t < VERIFY_T; ++t) {
-  for (int b = 0; b < BATCH_SIZE; ++b) {
-    result[t][b] += simd_shuffle_down(result[t][b], 8);
-    result[t][b] += simd_shuffle_down(result[t][b], 4);
-    result[t][b] += simd_shuffle_down(result[t][b], 2);
-    result[t][b] += simd_shuffle_down(result[t][b], 1);
-  }
-}
-
-if (k_lane == 0 && out_row < N_SIZE) {
-  for (int t = 0; t < VERIFY_T; ++t) {
-    for (int b = 0; b < BATCH_SIZE; ++b) {
-      y[(b * VERIFY_T + t) * N_SIZE + out_row] = T(result[t][b]);
-    }
-  }
-}
-"""
-
-
-_MXFP8_SINGLE_SOURCE = r"""
-uint tile = threadgroup_position_in_grid.y;
-uint simd_gid = simdgroup_index_in_threadgroup;
-uint simd_lid = thread_index_in_simdgroup;
-
-constexpr int RESULTS = 4;
-constexpr int ROWS_PER_TG = 2 * RESULTS;
-constexpr int VALUES_PER_THREAD = 8;
-constexpr int BLOCK_SIZE = 32 * VALUES_PER_THREAD;
-constexpr int GROUPS = K_SIZE / GROUP_SIZE;
-
-int out_row = int(tile) * ROWS_PER_TG + int(simd_gid) * RESULTS;
-float result[VERIFY_T][RESULTS];
-for (int t = 0; t < VERIFY_T; ++t) {
-  for (int row = 0; row < RESULTS; ++row) {
-    result[t][row] = 0.0f;
-  }
-}
-
-const device uint8_t* ws =
-    (const device uint8_t*)w + out_row * K_SIZE + int(simd_lid) * VALUES_PER_THREAD;
-const device uint8_t* ss =
-    scales + out_row * GROUPS + int(simd_lid) / (GROUP_SIZE / VALUES_PER_THREAD);
-
-for (int k = 0; k < K_SIZE; k += BLOCK_SIZE) {
-  float xt[VERIFY_T][VALUES_PER_THREAD];
-  for (int t = 0; t < VERIFY_T; ++t) {
-    const device T* source = x + t * K_SIZE + k + int(simd_lid) * VALUES_PER_THREAD;
-    for (int i = 0; i < VALUES_PER_THREAD; ++i) {
-      xt[t][i] = float(source[i]);
-    }
-  }
-  for (int row = 0; row < RESULTS; ++row) {
-    const device uint8_t* wr = ws + row * K_SIZE;
-    const device uint8_t* sr = ss + row * GROUPS;
-    float scale = decode_e8m0(sr[0]);
-    for (int t = 0; t < VERIFY_T; ++t) {
-      float accum = 0.0f;
-      for (int i = 0; i < VALUES_PER_THREAD; ++i) {
-        accum += xt[t][i] * decode_e4m3(wr[i]);
-      }
-      result[t][row] += scale * accum;
-    }
-  }
-  ws += BLOCK_SIZE;
-  ss += BLOCK_SIZE / GROUP_SIZE;
-}
-
-for (int t = 0; t < VERIFY_T; ++t) {
-  for (int row = 0; row < RESULTS; ++row) {
-    float value = simd_sum(result[t][row]);
-    if (simd_lid == 0 && out_row + row < N_SIZE) {
-      y[t * N_SIZE + out_row + row] = T(value);
-    }
   }
 }
 """
@@ -1335,42 +949,6 @@ def _hc_expand_kernel(dtype, rows, hc_mult, width):
 
 
 @lru_cache(maxsize=None)
-def _affine_wide_kernel(bits, dtype, batch, length, k_size, n_size, group_size):
-    return mx.fast.metal_kernel(
-        name=(
-            f"glm5_next_verify_affine{bits}_wide_"
-            f"{_dtype_name(dtype)}_b{batch}_t{length}_k{k_size}_n{n_size}_g{group_size}"
-        ),
-        input_names=["x", "w", "scales", "biases"],
-        output_names=["y"],
-        header=_AFFINE5_HEADER if bits == 5 else _AFFINE4_HEADER,
-        source=_affine_source(_AFFINE5_WIDE_SOURCE, bits),
-    )
-
-
-@lru_cache(maxsize=None)
-def _affine_wide_argmax_kernel(
-    bits,
-    dtype,
-    batch,
-    length,
-    k_size,
-    n_size,
-    group_size,
-):
-    return mx.fast.metal_kernel(
-        name=(
-            f"glm5_next_verify_affine{bits}_wide_argmax_"
-            f"{_dtype_name(dtype)}_b{batch}_t{length}_k{k_size}_n{n_size}_g{group_size}"
-        ),
-        input_names=["x", "w", "scales", "biases"],
-        output_names=["tile_values", "tile_indices"],
-        header=_AFFINE5_HEADER if bits == 5 else _AFFINE4_HEADER,
-        source=_affine_source(_AFFINE5_WIDE_ARGMAX_SOURCE, bits),
-    )
-
-
-@lru_cache(maxsize=None)
 def _affine_switch_gate_up_kernel(
     bits,
     dtype,
@@ -1399,29 +977,6 @@ def _affine_switch_gate_up_kernel(
         output_names=["up_y", "gate_y"],
         header=_affine_exact_header(bits, group_size, 1, 4),
         source=_affine_source(_AFFINE5_SWITCH_GATE_UP_SOURCE, bits),
-    )
-
-
-@lru_cache(maxsize=None)
-def _affine_dense_block_kernel(
-    bits,
-    dtype,
-    length,
-    k_size,
-    n_size,
-    group_size,
-    packs_per_thread,
-):
-    return mx.fast.metal_kernel(
-        name=(
-            f"glm5_next_verify_affine{bits}_dense_block_"
-            f"{_dtype_name(dtype)}_t{length}_k{k_size}_n{n_size}_"
-            f"g{group_size}_p{packs_per_thread}"
-        ),
-        input_names=["x", "w", "scales", "biases"],
-        output_names=["y"],
-        header=_affine_exact_header(bits, group_size, 1, 1, packs_per_thread),
-        source=_affine_source(_AFFINE5_DENSE_BLOCK_SOURCE, bits),
     )
 
 
@@ -1456,34 +1011,6 @@ def _affine_moe_down_kernel(
         output_names=["y"],
         header=_affine_exact_header(bits, group_size, 2, 4),
         source=_affine_source(_AFFINE5_MOE_DOWN_SOURCE, bits),
-    )
-
-
-@lru_cache(maxsize=None)
-def _mxfp8_wide_kernel(dtype, batch, length, k_size, n_size, group_size):
-    return mx.fast.metal_kernel(
-        name=(
-            "glm5_next_verify_mxfp8_wide_"
-            f"{_dtype_name(dtype)}_b{batch}_t{length}_k{k_size}_n{n_size}_g{group_size}"
-        ),
-        input_names=["x", "w", "scales"],
-        output_names=["y"],
-        header=_MXFP8_HEADER,
-        source=_MXFP8_WIDE_SOURCE,
-    )
-
-
-@lru_cache(maxsize=None)
-def _mxfp8_single_kernel(dtype, length, k_size, n_size, group_size):
-    return mx.fast.metal_kernel(
-        name=(
-            "glm5_next_verify_mxfp8_single_"
-            f"{_dtype_name(dtype)}_t{length}_k{k_size}_n{n_size}_g{group_size}"
-        ),
-        input_names=["x", "w", "scales"],
-        output_names=["y"],
-        header=_MXFP8_HEADER,
-        source=_MXFP8_SINGLE_SOURCE,
     )
 
 
@@ -1529,124 +1056,6 @@ def exact_dense_block_linear(linear, x: mx.array) -> Optional[mx.array]:
         output_shapes=[(batch, length, out_size)],
         output_dtypes=[x.dtype],
     )[0]
-
-
-def exact_quantized_block_linear(linear, x: mx.array) -> Optional[mx.array]:
-    """Decode-exact BxT projection with weight reuse across verifier time."""
-    if (
-        not mx.metal.is_available()
-        or not isinstance(linear, nn.QuantizedLinear)
-        or x.ndim != 3
-        or not 1 < x.shape[1] <= 4
-        or not 1 <= x.shape[0] <= 4
-        or x.dtype not in (mx.bfloat16, mx.float16)
-    ):
-        return None
-
-    batch, length, k_size = x.shape
-    n_size = linear.weight.shape[0]
-    if k_size % linear.group_size or n_size < 8:
-        return None
-
-    x = mx.contiguous(x)
-    template = [
-        ("T", x.dtype),
-        ("BATCH_SIZE", int(batch)),
-        ("VERIFY_T", int(length)),
-        ("K_SIZE", int(k_size)),
-        ("N_SIZE", int(n_size)),
-        ("GROUP_SIZE", int(linear.group_size)),
-    ]
-
-    if (
-        linear.mode == "affine"
-        and linear.bits in (4, 5)
-        and linear.biases is not None
-        and linear.scales.dtype == x.dtype
-        and linear.biases.dtype == x.dtype
-        and linear.group_size % 8 == 0
-    ):
-        if batch == 1 and not (k_size % 512 or n_size % 16):
-            packs_per_thread = 2
-            kernel = _affine_dense_block_kernel(
-                linear.bits,
-                x.dtype,
-                length,
-                k_size,
-                n_size,
-                linear.group_size,
-                packs_per_thread,
-            )
-            output = kernel(
-                inputs=[x, linear.weight, linear.scales, linear.biases],
-                template=[
-                    ("T", x.dtype),
-                    ("VERIFY_T", int(length)),
-                    ("K_SIZE", int(k_size)),
-                    ("N_SIZE", int(n_size)),
-                    ("GROUP_SIZE", int(linear.group_size)),
-                ],
-                grid=(32, 16 * (n_size // 16), 1),
-                threadgroup=(32, 1, 1),
-                output_shapes=[(batch, length, n_size)],
-                output_dtypes=[x.dtype],
-            )[0]
-            if "bias" in linear:
-                output = output + linear["bias"]
-            return output
-        if batch == 1:
-            return None
-        # MLX uses a different short-vector reduction for decode-shaped
-        # Bx1 affine projections. The wide verifier kernel follows the regular
-        # group reduction and is therefore not bit exact below 512 inputs.
-        # Let the verifier's batched MLX fallback preserve the B-row kernel.
-        if k_size < 512:
-            return None
-        kernel = _affine_wide_kernel(
-            linear.bits,
-            x.dtype,
-            batch,
-            length,
-            k_size,
-            n_size,
-            linear.group_size,
-        )
-        inputs = [x, linear.weight, linear.scales, linear.biases]
-        rows_per_tg = 8
-    elif (
-        linear.mode == "mxfp8"
-        and linear.bits == 8
-        and linear.group_size == 32
-        and linear.biases is None
-        and linear.scales.dtype == mx.uint8
-    ):
-        if batch == 1:
-            if k_size % 512 or n_size % 8:
-                return None
-            kernel = _mxfp8_single_kernel(
-                x.dtype, length, k_size, n_size, linear.group_size
-            )
-            rows_per_tg = 8
-        else:
-            kernel = _mxfp8_wide_kernel(
-                x.dtype, batch, length, k_size, n_size, linear.group_size
-            )
-            rows_per_tg = 4
-        inputs = [x, linear.weight, linear.scales]
-    else:
-        return None
-
-    output = kernel(
-        inputs=inputs,
-        template=template,
-        grid=(32, 2 * ((n_size + rows_per_tg - 1) // rows_per_tg), 1),
-        threadgroup=(32, 2, 1),
-        output_shapes=[(batch, length, n_size)],
-        output_dtypes=[x.dtype],
-    )[0]
-    if "bias" in linear:
-        output = output + linear["bias"]
-    return output
 
 
 def exact_affine_switch_gate_up(switch, x: mx.array, indices: mx.array):
@@ -2003,94 +1412,6 @@ def exact_hc_expand(
         output_shapes=[residual.shape],
         output_dtypes=[x.dtype],
     )[0]
-    return output
-
-
-def exact_quantized_block_argmax(linear, x: mx.array) -> Optional[mx.array]:
-    """Greedy affine projection without materializing vocabulary logits."""
-    if (
-        not mx.metal.is_available()
-        or not isinstance(linear, nn.QuantizedLinear)
-        or linear.mode != "affine"
-        or linear.bits not in (4, 5)
-        or linear.biases is None
-        or "bias" in linear
-        or x.ndim != 3
-        or not 2 <= x.shape[0] <= 4
-        or not 1 < x.shape[1] <= 4
-        or x.dtype not in (mx.bfloat16, mx.float16)
-        or linear.scales.dtype != x.dtype
-        or linear.biases.dtype != x.dtype
-    ):
-        return None
-
-    batch, length, k_size = x.shape
-    n_size = linear.weight.shape[0]
-    if k_size % linear.group_size or linear.group_size % 8 or n_size < 8 or n_size % 8:
-        return None
-
-    num_tiles = n_size // 8
-    x = mx.contiguous(x)
-    kernel = _affine_wide_argmax_kernel(
-        linear.bits,
-        x.dtype,
-        batch,
-        length,
-        k_size,
-        n_size,
-        linear.group_size,
-    )
-    tile_values, tile_indices = kernel(
-        inputs=[x, linear.weight, linear.scales, linear.biases],
-        template=[
-            ("T", x.dtype),
-            ("BATCH_SIZE", int(batch)),
-            ("VERIFY_T", int(length)),
-            ("K_SIZE", int(k_size)),
-            ("N_SIZE", int(n_size)),
-            ("GROUP_SIZE", int(linear.group_size)),
-            ("NUM_TILES", int(num_tiles)),
-        ],
-        grid=(32, 2 * num_tiles, 1),
-        threadgroup=(32, 2, 1),
-        output_shapes=[
-            (batch, length, num_tiles),
-            (batch, length, num_tiles),
-        ],
-        output_dtypes=[x.dtype, mx.int32],
-    )
-    best_tile = mx.argmax(tile_values, axis=-1)
-    return mx.take_along_axis(
-        tile_indices,
-        best_tile[..., None],
-        axis=-1,
-    ).squeeze(-1)
-
-
-def batched_time_quantized_linear(linear, x: mx.array) -> Optional[mx.array]:
-    """Project verifier positions while retaining decode's batch dimension."""
-    if not isinstance(linear, nn.QuantizedLinear) or x.ndim != 3 or x.shape[1] <= 1:
-        return None
-
-    length = x.shape[1]
-    transposed = mx.contiguous(x.transpose(1, 0, 2))
-    weight = mx.broadcast_to(linear.weight[None], (length, *linear.weight.shape))
-    scales = mx.broadcast_to(linear.scales[None], (length, *linear.scales.shape))
-    biases = linear.get("biases")
-    if biases is not None:
-        biases = mx.broadcast_to(biases[None], (length, *biases.shape))
-    output = mx.quantized_matmul(
-        transposed,
-        weight,
-        scales=scales,
-        biases=biases,
-        transpose=True,
-        group_size=linear.group_size,
-        bits=linear.bits,
-        mode=linear.mode,
-    ).transpose(1, 0, 2)
-    if "bias" in linear:
-        output = output + linear["bias"]
     return output
