+++ mlx_vlm/models/nemotron_h/speculative_verifier.py
@@ -103,190 +103,6 @@ _MAMBA_STATE_SOURCE = r"""
     }
 """
 
-_NVFP4_ARGMAX_HEADER = r"""
-    using namespace metal;
-
-    constant constexpr int SIMD_SIZE = 32;
-    constant constexpr int VALUES_PER_THREAD = 8;
-    constant constexpr int BLOCK_SIZE = SIMD_SIZE * VALUES_PER_THREAD;
-    constant constexpr int RESULTS_PER_SIMDGROUP = 4;
-    constant constexpr int NUM_SIMDGROUPS = 2;
-    constant constexpr int ROWS_PER_THREADGROUP =
-        RESULTS_PER_SIMDGROUP * NUM_SIMDGROUPS;
-
-    inline float decode_e2m1(uint value) {
-        half decoded = as_type<half>(ushort((value & 7) << 9));
-        decoded *= 16384.0h;
-        return float(value & 8 ? -decoded : decoded);
-    }
-
-    inline float decode_e4m3(uint8_t value) {
-        half decoded = as_type<half>(ushort((value & 127) << 7));
-        decoded *= 256.0h;
-        return float(value & 128 ? -decoded : decoded);
-    }
-
-    inline float nvfp4_dot(
-        const device uint8_t* weight,
-        const thread float* values,
-        float scale) {
-        const device uint16_t* packed =
-            reinterpret_cast<const device uint16_t*>(weight);
-        float result = 0.0f;
-        for (int index = 0; index < 2; ++index) {
-            uint word = packed[index];
-            result +=
-                values[4 * index] * decode_e2m1(word) +
-                values[4 * index + 1] * decode_e2m1(word >> 4) +
-                values[4 * index + 2] * decode_e2m1(word >> 8) +
-                values[4 * index + 3] * decode_e2m1(word >> 12);
-        }
-        return scale * result;
-    }
-"""
-
-_NVFP4_ARGMAX_SOURCE = r"""
-    uint output_tile = threadgroup_position_in_grid.y;
-    uint batch_index = threadgroup_position_in_grid.z;
-    uint simd_group = simdgroup_index_in_threadgroup;
-    uint lane = thread_index_in_simdgroup;
-    uint output_row =
-        output_tile * ROWS_PER_THREADGROUP +
-        simd_group * RESULTS_PER_SIMDGROUP;
-
-    threadgroup float group_values[VERIFY_T][NUM_SIMDGROUPS];
-    threadgroup int group_indices[VERIFY_T][NUM_SIMDGROUPS];
-    float results[VERIFY_T][RESULTS_PER_SIMDGROUP];
-    float values[VERIFY_T][VALUES_PER_THREAD];
-    for (int token = 0; token < VERIFY_T; ++token) {
-        for (int row = 0; row < RESULTS_PER_SIMDGROUP; ++row) {
-            results[token][row] = 0.0f;
-        }
-    }
-
-    const device uint8_t* weight_bytes =
-        reinterpret_cast<const device uint8_t*>(w);
-    for (int k = 0; k < K_SIZE; k += BLOCK_SIZE) {
-        int lane_start = k + int(lane) * VALUES_PER_THREAD;
-        if (lane_start < K_SIZE) {
-            for (int token = 0; token < VERIFY_T; ++token) {
-                const device T* input =
-                    x + (int(batch_index) * VERIFY_T + token) * K_SIZE + lane_start;
-                for (int index = 0; index < VALUES_PER_THREAD; ++index) {
-                    values[token][index] = float(input[index]);
-                }
-            }
-
-            for (int row = 0; row < RESULTS_PER_SIMDGROUP; ++row) {
-                int n = int(output_row) + row;
-                const device uint8_t* weight =
-                    weight_bytes + n * (K_SIZE / 2) + lane_start / 2;
-                float scale = decode_e4m3(
-                    scales[n * (K_SIZE / 16) + lane_start / 16]
-                );
-                for (int token = 0; token < VERIFY_T; ++token) {
-                    results[token][row] += nvfp4_dot(weight, values[token], scale);
-                }
-            }
-        }
-    }
-
-    for (int token = 0; token < VERIFY_T; ++token) {
-        float best_value = -3.4028234663852886e38f;
-        int best_index = 0;
-        for (int row = 0; row < RESULTS_PER_SIMDGROUP; ++row) {
-            int n = int(output_row) + row;
-            float value = float(T(simd_sum(results[token][row])));
-            if (value > best_value) {
-                best_value = value;
-                best_index = n;
-            }
-        }
-        if (lane == 0) {
-            group_values[token][simd_group] = best_value;
-            group_indices[token][simd_group] = best_index;
-        }
-    }
-    threadgroup_barrier(mem_flags::mem_threadgroup);
-
-    if (simd_group == 0 && lane == 0) {
-        for (int token = 0; token < VERIFY_T; ++token) {
-            float best_value = group_values[token][0];
-            int best_index = group_indices[token][0];
-            for (int group = 1; group < NUM_SIMDGROUPS; ++group) {
-                float value = group_values[token][group];
-                if (value > best_value) {
-                    best_value = value;
-                    best_index = group_indices[token][group];
-                }
-            }
-            int offset =
-                (int(batch_index) * VERIFY_T + token) * NUM_TILES +
-                int(output_tile);
-            tile_values[offset] = T(best_value);
-            tile_indices[offset] = best_index;
-        }
-    }
-"""
-
-
-@lru_cache(maxsize=None)
-def _nvfp4_argmax_kernel(dtype, verify_length, input_size, output_size):
-    dtype_name = {mx.bfloat16: "bf16", mx.float16: "fp16"}.get(dtype, "unk")
-    return mx.fast.metal_kernel(
-        name=(
-            "nemotron_h_nvfp4_argmax_"
-            f"t{verify_length}_k{input_size}_n{output_size}_{dtype_name}"
-        ),
-        input_names=["x", "w", "scales"],
-        output_names=["tile_values", "tile_indices"],
-        header=_NVFP4_ARGMAX_HEADER,
-        source=_NVFP4_ARGMAX_SOURCE,
-    )
-
-
-def _nvfp4_argmax(linear, hidden: mx.array) -> Optional[mx.array]:
-    if (
-        not isinstance(linear, nn.QuantizedLinear)
-        or linear.mode != "nvfp4"
-        or linear.bits != 4
-        or linear.group_size != 16
-        or "bias" in linear
-        or hidden.ndim != 3
-        or hidden.dtype not in (mx.bfloat16, mx.float16)
-        or not mx.metal.is_available()
-        or mx.default_device() != mx.gpu
-    ):
-        return None
-
-    batch, length, input_size = hidden.shape
-    expected_input_size = linear.weight.shape[1] * 8
-    output_size = linear.weight.shape[0]
-    if input_size != expected_input_size or input_size % 16 or output_size % 8:
-        return None
-
-    tile_count = output_size // 8
-    kernel = _nvfp4_argmax_kernel(hidden.dtype, length, input_size, output_size)
-    tile_values, tile_indices = kernel(
-        inputs=[mx.contiguous(hidden), linear.weight, linear.scales],
-        template=[
-            ("T", hidden.dtype),
-            ("VERIFY_T", int(length)),
-            ("K_SIZE", int(input_size)),
-            ("N_SIZE", int(output_size)),
-            ("NUM_TILES", int(tile_count)),
-        ],
-        grid=(32, 2 * tile_count, batch),
-        threadgroup=(32, 2, 1),
-        output_shapes=[
-            (batch, length, tile_count),
-            (batch, length, tile_count),
-        ],
-        output_dtypes=[hidden.dtype, mx.int32],
-    )
-    best_tile = mx.argmax(tile_values, axis=-1)
-    return mx.take_along_axis(tile_indices, best_tile[..., None], axis=-1).squeeze(-1)
-
 
 @lru_cache(maxsize=None)
 def _mamba_verify_kernel(
@@ -494,17 +310,6 @@ def _mamba_update_exact(
 
 class NemotronHExactSpeculativeVerifier(Qwen3_5ExactSpeculativeVerifier):
     """Run Nemotron-H verification with singleton-equivalent Metal kernels."""
-
-    def quantized_argmax(
-        self,
-        linear,
-        hidden: mx.array,
-        token_mask: Optional[mx.array] = None,
-    ) -> Optional[mx.array]:
-        output = super().quantized_argmax(linear, hidden, token_mask=token_mask)
-        if output is not None or token_mask is not None:
-            return output
-        return _nvfp4_argmax(linear, hidden)
 
     def _linear(self, linear, hidden: mx.array) -> mx.array:
         if hidden.ndim == 3 and hidden.shape[1] > 1 and hidden.dtype == mx.float32:
